+++ sample/objective-c/GNAdSampleAdMobAdapter/GNAdSampleAdMobAdapter/ViewController.h
@@ -1,0 +1,14 @@
+//
+//  ViewController.h
+//  GNAdSampleAdMobAdapter
+//
+
+#import <UIKit/UIKit.h>
+@class GADBannerView;
+@class GADBannerViewDelegate;
+
+@interface ViewController : UIViewController<GADBannerViewDelegate> {
+    GADBannerView *_bannerView;
+}
+
+@end