+++ sample/objective-c/GNAdSampleBanner/GNAdSampleBanner/ViewController.h
@@ -1,0 +1,15 @@
+//
+//  ViewController.h
+//  GNAdSampleBanner
+//
+
+#import <UIKit/UIKit.h>
+#import <GNAdSDK/GNAdView.h>
+
+@interface ViewController : UIViewController<GNAdViewDelegate>
+{
+    GNAdView *_adView;
+}
+
+@end
+