+++ sample/objective-c/GNAdSampleInterstitial/GNAdSampleInterstitial/ViewController.h
@@ -1,0 +1,15 @@
+//
+//  ViewController.h
+//  GNAdSampleInterstitial
+//
+
+#import <UIKit/UIKit.h>
+#import <GNAdSDK/GNInterstitial.h>
+
+@interface ViewController : UIViewController<GNInterstitialDelegate>
+{
+    GNInterstitial *_interstitial;
+}
+
+@end
+