+++ sample/objective-c/GNAdSampleMultipleBanner/GNAdSampleMultipleBanner/TableViewCell.m
@@ -1,0 +1,19 @@
+//
+//  TableViewCell.m
+//  GNAdSampleMultipleBanner
+//
+
+#import "TableViewCell.h"
+
+@interface TableViewCell()
+
+@end
+
+@implementation TableViewCell
+
+@synthesize icon;
+@synthesize title;
+@synthesize description;
+@synthesize adView;
+
+@end