+++ sample/objective-c/GNAdSampleVideo/GNAdSampleVideo/ViewController.h
@@ -1,0 +1,15 @@
+//
+//  ViewController.h
+//  GNAdSampleVideo
+//
+
+#import <UIKit/UIKit.h>
+#import <GNAdSDK/GNAdVideo.h>
+
+@interface ViewController : UIViewController<GNAdVideoDelegate>
+{
+    GNAdVideo *_videoAd;
+}
+
+@end
+