+++ sample/swift/GNAdSampleMultipleBanner/TableViewCell.swift
@@ -1,0 +1,23 @@
+//
+//  TableViewCell.swift
+//  GNAdSampleMultipleBanner
+//
+
+import UIKit
+
+class TableViewCell: UITableViewCell {
+
+    @IBOutlet weak var icon: UIImageView!
+    @IBOutlet weak var title: UILabel!
+    @IBOutlet weak var content: UILabel!
+    weak var adView: GNAdView!
+
+    override func awakeFromNib() {
+        super.awakeFromNib()
+    }
+
+    override func setSelected(selected: Bool, animated: Bool) {
+        super.setSelected(selected, animated: animated)
+    }
+
+}