+++ CHANGELOG.md
@@ -1,10 +1,32 @@
-## Version 2.0.7
+## Version 2.1.4
 
-- **Genieeメディエーション機能の追加**.
-    - iAd広告の配信
-    - AdMob広告の配信
+- **Add Natvie Ad click tracking**
+
+## Version 2.1.3
+
+- **Add Bitcode for iOS9**
+
+## Version 2.1.2
+
+- Released SDK as a framework.
+- Improved banner ad loading time.
+- Improved interstitial ad loading time.
+
+## Version 2.1.1
+
+- Add Multiple banner and custom size
+
+## Version 2.0.9
+
+- add getting multiple native ad
 
 ## Version 2.0.8
 
 - **Native Ad 機能の追加**.
-    - In-Feed 広告の配信
+    - In-Feed Native Ad 広告の配信
+
+## Version 2.0.7
+
+- **Geniee SDK 経由 iAd、Admob 広告の配信**.
+    - iAd広告の配信
+    - AdMob広告の配信